Патч -Сократил время учёта статистикой пользователя с 5 до 1 минуты; -Обновил логику подсчёта кол-ва пользователей по сайту и страницам: теперь считается целое число людей среди ненулевых значений.

+++ api/api.v1.stats/DTOs/Attendance/AttendanceQuantityPlotDTO.cs
@@ -1,4 +1,4 @@
 ﻿namespace api.v1.stats.DTOs.Attendance
 {
-    public sealed record AttendanceQuantityPlotDTO(double LeftDate, double RightDate, double Quantity);
+    public sealed record AttendanceQuantityPlotDTO(double LeftDate, double RightDate, int Quantity);
 }
